Renames bash_profile to bashrc, adds install script
New bash_profile that sources bashrc. This is to enable easier
transferral and of dotfiles between Linux and OS X. This is because by
default terminal on OS X's Terminal.app (and iTerm 2) default to opening
a login shell (which sources bash_profile) and Linux's terminal emulator
defaults to using an interactive shell (requiring bashrc) when in the
GUI and a login shell when SSHing in. Now everything is the same. Yey!